+++ srcs/objects/sphere.c
@@ -1,21 +1,21 @@
 #include "miniRT.h"
 
-float	sphere_intersection(t_minirt *minirt, t_ray ray, t_object sphere)
+double	sphere_intersection(t_minirt *minirt, t_ray ray, t_object sphere)
 {
 	(void)minirt;
 	/***** a voir  *****/
 	t_vec3 offset_ray_origin = sub_vec3(ray.origin, sphere.position); // vecteur entre le point d'origine et le centre de la sphere
 
 
-	float a = dot_vec3(ray.direction, ray.direction); 
-	float b = 2.0f * dot_vec3(offset_ray_origin, ray.direction); // ou off set ray origin / ray.origin ?
-	float c = dot_vec3(offset_ray_origin, offset_ray_origin) - sphere.spec.sphere.radius * sphere.spec.sphere.radius;
-	float discriminant = b * b - 4.0f * a * c;
+	double a = dot_vec3(ray.direction, ray.direction); 
+	double b = 2.0f * dot_vec3(offset_ray_origin, ray.direction); // ou off set ray origin / ray.origin ?
+	double c = dot_vec3(offset_ray_origin, offset_ray_origin) - sphere.spec.sphere.radius * sphere.spec.sphere.radius;
+	double discriminant = b * b - 4.0f * a * c;
 	if (discriminant < 0.f) 
 		return (-1);
 	// colorisation de la scene
-	float t0 = (-b - sqrt(discriminant)) / (2.0f * a); // calcul de la distance entre le point d'origine et le point d'intersection
-	float t1 = (-b + sqrt(discriminant)) / (2.0f * a);
+	double t0 = (-b - sqrt(discriminant)) / (2.0f * a); // calcul de la distance entre le point d'origine et le point d'intersection
+	double t1 = (-b + sqrt(discriminant)) / (2.0f * a);
 	if (t0 < 0) // je checke si je suis dans la sphere et je prends la distance la plus proche
 		t0 = t1;
 	if (t0 < 0) // je checke si l'objet n'est pas derriere la camera
