+++ srcs/vector/dot_product.c
@@ -14,7 +14,7 @@ and a zero dot product indicates that the vectors are perpendicular.
 A⋅B = A1​B1 ​+ A2​B2 ​+ A3​B3
 */
 
-float	dot_vec3(t_vec3 a, t_vec3 b)
+double	dot_vec3(t_vec3 a, t_vec3 b)
 {
 	return (a.x * b.x + a.y * b.y + a.z * b.z);
 }
